# gateway/gen/openapiv2/auth.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "option go_package = \"github.com/anshulrgoyal/grpc-gateway-demo;grpc_gateway_demo\";",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Biz"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/get_users": {
      "get": {
        "operationId": "Biz_get_users",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authget_user_output"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "authKey",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "messageId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Biz"
        ]
      }
    },
    "/get_users_sql": {
      "get": {
        "operationId": "Biz_get_users_with_sql_injection",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authget_user_output"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "authKey",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "messageId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Biz"
        ]
      }
    },
    "/req_dh_params": {
      "get": {
        "operationId": "Auth_req_DH_params",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authReq_DH_params_response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "nonce",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "serverNonce",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "messageId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "a",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/req_pq": {
      "get": {
        "operationId": "Auth_req_pq",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authReq_pq_response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "nonce",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "messageId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "definitions": {
    "authReq_DH_params_response": {
      "type": "object",
      "properties": {
        "nonce": {
          "type": "string"
        },
        "serverNonce": {
          "type": "string"
        },
        "messageId": {
          "type": "integer",
          "format": "int32"
        },
        "b": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "authReq_pq_response": {
      "type": "object",
      "properties": {
        "nonce": {
          "type": "string"
        },
        "serverNonce": {
          "type": "string"
        },
        "messageId": {
          "type": "integer",
          "format": "int32"
        },
        "p": {
          "type": "integer",
          "format": "int32"
        },
        "g": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "authUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "family": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "sex": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "authget_user_output": {
      "type": "object",
      "properties": {
        "user": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/authUser"
          }
        },
        "messageId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}